Finally reworked crazyGravity
Been sitting on this for a while, it does the following:
When Mario's height goes above mLastGroundedPos.y + 7000, Mario's gravity gets set back to default(1.0). Because of something that resembles inertia, this doesn't mean Mario's height is hard capped to mLastGroundedPos.y + 7000, as he'll carry some of that momentum up even when gravity goes back to 1.0.